feat(web): reuse the family-invite screen in a /design study
Extract the /family/accept invite content into a shared FamilyInviteScreen
component so the live join page and the design showcase render identical
copy and CTA branching from one source. Add a "Family plan invite
acceptance" study under /design (Sections) covering the happy path plus
the terminal states, using synthetic preview data only.

Also tighten the invite CTA copy (drop the redundant Messages helper line,
shorten the web sign-in helper, and reword the secondary sign-in link).
